refactor(SRender): start adaptors from appConfig.
The adaptors are no longer started and stopped by the render service. To achieve that, I also had to change the way we create the render window. Now I create a render window with a dummy size of 1x1 in the constructor of the widget. Thus I no longer have to wait for the widget to be painted in order to create the render window. This simplifies a lot the start process.
Last I modified some code in the light adaptors creation.