fix(AppConfigManager): swap correctly a key if the object is different
We forgot to handle the case when a deferred object is replaced by a different instance. We used to think there will always be a add/remove, but if several services produce the same object, we may have instead add/add/add/.../remove.
So we simply test if the registered object is different than the new one, and if yes, we swap the key.