fix(fwAtomsBoostIO): do not trust boost ptree ordering
When reading a sequence, we add all elemnents of the property tree to
the sequence in the order supplied by boost's property_tree. This is not
good. According to the docs;

"It is very important to remember that the property sequence is
not ordered by the key. It preserves the order of insertion."

The fix is to do the sort ourself based on the keys of the subatoms.