fix(IService): allow slots exceptions to be catched from another worker.
Catching exceptions for start(), stop(), swap(), swapKey() and update()
was broken since I allowed services to be reused after triggering an
exception. It worked well when ran on the same worker, but it was broken
when ran on another thread.
This commit fixes this by rethrowing the exception when we call one of
these methods asynchronously.