fix(Mesh): switch to normalized unsigned bytes for vertex color format
I had to introduce this hack years ago because of a bug in the RenderToVertexBuffer code, but it seems the bug has been fixed upstream since everything is now alright.
So now we use the regular vertex color format, and thus our shaders are compatible with ManualObjects with vertex colors.