fix(fwTest): remove broken usage of fwTest::Thread::timedJoin()
When removing global usage of ::boost::thread in our code, we misreplaced the code of this function. It was designed to timeout if a thread failed to join. The new code first wait the required time, then join. Thus the function could never timeout and on top of that it slowed down all the tests using this function since it always waited. This was rewritten using std::future and thus the whole class is now useless and has been removed.