feat(IService): discard update while stopped, remove assertion
We use to assert that an update is not performed after the service is stopped. However, this may happens sometimes when services do not run on the same worker, an update request may still lie in the event loop after the stop. We used to cheat by using a different slot where the assertion is not raised. I propose here to discard the update and just raise a warning instead.