fix(SFrameGrabber+SMatricesReader): current index position in setStep
Calling setStep slots changed the current step but not the current index position. We need to also update this indices position because it's current value is update right after reading a next image/matrix. If this indices are not changed setStep slots, the step will be take into account only on the next call of readNext slots.