fix(macros.cmake): remove version in the bundle library file name
It is not necessary to have the version number in the bundle library name.  On Linux, it is already present after the .so extension, thus this led to get it twice. We do not have it on Windows but anyway we got the version number in the bundle folder just in case. We already removed them from the regular libraries a while ago.