chore(dcmtk): remove copies of dicom.dic and storescp.cfg
Instead of copying those files in our source tree, we copy them at build time. This avoids to get licence error reports with FOSSA and we are sure that these are the raw files and not modified version. And also they are in sync with the current version of DCMTK.