feat(fwGuiQml): create FileDialog QtQuick popup when use of LocationDialog
Need to put a verification inside SImageSeries because even if it is modal, it doesn't block the flow of the App background. Need to create createComponent function inside QmlEngine to retrieve the QObject * that just has been created and with it invoke the method open of the FileDialog inside the App.
chore(fwGuiQml): erase QApplication and replace it by QGuiApplication

QWidget does not work with QGuiApplication and we need to erase all include of QApplication to change for the new App type and change activeWindow() by focusWindow()