fix(WorkerAsio): avoid race conditions
When destroying a worker, we may ask it to join twice the thread if someone else is calling stop(). On top of that, hen stopping a worker, we decide whether we need to join the thread by testing the state of the io_service. However this does not seem to be a reliable way, so we now test the thread state itself.

This solves a lot of random errors when lauching all unit-tests.