fix(fwDataTest): fix MTLockTest failure
The test try to lock the data to write "lili" or "tata" in a string data,  it launch two asynchronous methods that lock the data to write, wait 10ms between each char insertion and wait 5ms after the unlock. Then it ensures that the letters from "tata" and "lili" are not mixed in the string.