fix(fwGuiQml): call endResetModel protected function to notify data change
Change the order of function call for the model by setting first the context to erase the Context not found warning. Now we set it at the beginning by calling endResetModel everytime we add data.
chore(fwGuiQml): delete useless file to make qml work

All those files that are erase are not used anywhere
chore(guiQml): change the Required bundle from guiQt to guiQml

Change it on each qml tutorial