fix(PoCDialog): change enable state in each example where it can throw an assert
Each open button where it can throw an assert has been disable and inside the progressdialog it's the add button which is disabled and changed order of list inside multiselectordialog because each option is not add in the good order
chore(PoCDialog): remove useless dependance and file and move the qmlRegister inside plugin

Remove the useless RadioButton inside logger and remove the baseline inside the text which does some visual bug. And remove dependance inside guiQml.
chore(fwQml): remove function never called and the string filepath

The filepath is not used after initialization and createComponent was used just at the beginning of this issue.
fix(fwGuiQml): change the order of push inside the model from push front to push back

Before the model added in inverse order and i needed to change the loop order when giving result. Now it's good.
chore(guiQml): add plugins for future

Add some Qml plugins for the future