feat(SActivitySequencer): add new editor to display an activity stepper
The list of the  activities are displayed in a toolbar, the buttons are enabled only if the activity can be launched (if all its parameters are present). The sequencer uses the data from the previously launched activity to create the current one.

Create a new 'ActivityLauncher' configuration to simplify the launch of an activity sequencer.

Improve sample to use 'ActivitySequencer' configuration:
- use 'ActivitySequencer' config
- add volume rendering activity