fix(Image): fix the iterator when using a different type than the image
The final iterator is now computed based on the format of the iterator instead of using the image type.

You can now use Iterator<char> on int16 image. It can be useful to fill the image with zeros. You can also parse int32 image with int64 iterator to gain performance.