fix(AppConfig): fix XML variable regex
The regex is used to know if the xml attribute contains a variable (`${...}`). But in a few cases, for example in signal/slot connections, the variable is followed by other chars. For these cases, the regex should not be limited to `${...}`but to  `${...}....`.
The unit test are updated to check this error.