refactor(CMake): add intermediate object library to build shared libraries
Testing functions not exported from a shared library was impossible. With the will to hide implementation from public API, it would be a shame to lose the ability to test the implementation itself.
To overcome this, I introduced an intermediate object library. This way, the unit-test can link against it instead of the shared library. This way, it gains access to all the symbols, and thus the implementation.