fix(fwRuntime): enable automatically requirements
I tried to prevent to expose Module::setEnable() in the API, doing so I had to remove one use in fwServicesTest. Instead I used loadModule() but this function only loads enabled module. I don't really understand the reason, since on the other hand it starts automatically the depending module. So I think it is safe to recursively enable depending module when loading a module.