refactor(graphics): use an ogre listener to monitor camera movements
Previously, we triggered a custom signal from Layer, and caught in `SCamera`. This signal was triggered on any interaction in the window, even when no movement occurs. To fix this, to make the design simpler and to be able to avoid infinite loops when connecting services and data like in Tuto04SignalSlot, an Ogre listener has been installed in `SCamera`.