enh(viz): improve flat shading by using fragment shader computed normals
Flat shading may often be broken if point normals are shared between faces. To overcome this, we chose to compute normals in the fragment shader using eye position derivatives.
Doing so, the outdated and never used Gouraud shading mode was removed. This removes a shader combination which makes maintenance easier.