fix(viz): replace getTransformNode by a pure getter
`getTransformNode` was ambiguous since it return the transform node if existing or creates it if it doesn't.

This behaviour is kept but function was rennamed into `getOrCreateTransformNode`.
A pure getter was also created `getTransformNode()` to return the transform node pointer or nullptr if it doesn't exists.

Add small unit test to test ITransformable functions

All adaptors was updated to call either `getOrCreateTransformNode` or `getTransformNode` regarding the case (creation of the node of getting the pointer). This fixes the crash due to Ogre exception "Node already exists" when trying to get a deleted node.