fix(core): set the log output in the main of unit-tests
The redirection of the log into `fwTest.log` is performed by a singleton in sight::core::utest. We link with this library in coreTest, but we do not use any symbol from it. Thus the linker strips it and so in the end, we do not initialize the log output.
To fix this, we simply initialize it in the main file of unit tests.