enh(test): improve geometry::data::Mesh::transform unit-test
The unit test for `geometry::data::Mesh::transform` was pretty weak, as the test was mostly the same code as the tested code. The method is now tested with a hardcoded small mesh which is transformed, with all combinations of types possible (only points, point and point normals, point and cell normals, point and all normals).

Closes #80