refactor(viz): simplify transfer function management
This is a major rework of the transfer function data. Now the transfer function is a real piecewise function. Each piece stored in the object can be individually edited and thus is serialized. Pieces can be accessed and edited independently thanks to the `pieces()` function that returns a vector of `TransferFunctionData`. When the pieces are modified, the function `mergePieces()` must be called to mix the pieces together.

This implies that the duplicated `sight::module::ui::qt::SMultipleTF` and `sight::module::viz::scene2d::SMultipleTF` are removed in favor of `sight::module::ui::qt::STransferFunction` and `sight::module::viz::scene2d::STransferFunction`.

On top of that, the former TF "pool" is no longer exposed anywhere in the services interfaces, thus from the XML. `sight::module::ui::qt::STransferFunction` handles this "pool", renamed as "presets", internally.

Other changes were brought:
- `sight::data::TransferFunction::mergePieces()` implies the removal of `sight::module::data::SMergeTF` service. It also brings a functional change, the pieces of the transfer function is performed with an average of colors instead of a strict additive blending.
- It is no longer possible to move a transfer function piece outside the range of the viewport (bug fix request in the comments of this ticket).
- Adjusting the window in the negatoscope is now reflected properly in the transfer function editor (partially fixes https://git.ircad.fr/sight/sight/-/issues/847)
- Moved and renamed `TransferFunctionManager` from `scene3d` to `scene2D::TransferFunctionWindow`.

Closes #877