fix(io): do not restore the UUID of objects when deserializing
Restoring the UUID of objects implies that a serialized object is really unique. Thus if you read a file containing a object twice, you expect to have only on object in memory.
This is definitely not the behavior we want, we expect the object to be duplicated each time it is read, with its own UUID.