feat(viz): introduce GPU resource sharing
GPU resource sharing is introduced, with the first application to textures and transfer functions.
From the user point of view, we provide a new class `sight::viz::scene3d::Texture` and we extended the existing class `sight::viz::scene3d::TransferFunction` in order to allow GPU resource sharing. These classes only allocate Ogre resources (and thus GPU memory) when the Id of the source object changes. In addition, they "cache" the update of these resources. Thus now, when we have, for instance, 3 negato adaptors on the same image, there is only one `Ogre::Texture` created, and only one `Ogre::Texture` updated when the `sight::data::Image` is modified (and similarly for the transfer functions).