enh(viz): adapt default transfer function to a reference image
The default transfer function can now be adapted to the image pixel type, especially for ultrasound images.

The `data::Image` no longer holds a default transfer function. To get a default transfer function, you have to use `sight::module::ui::qt::image::STransferFunction` and specify a reference image as input. This is no issue if you don't want the transfer function selector in the user interface. You can register the service in an invisible view. You can have a look at `Tuto07GenericScene` for instance.

Also, some bugs were fixed in SNegato2DCamera for the computation of the viewport of the axial view.